ingestion/merge_to_bigquery/requirements.txt
google-cloud-bigquery==3.41.0
dotenv==0.9.9